refactor(examples): fold mcp-agent into demo-bank as a script
A whole workspace for one 207-line script, ~43% of which was a third copy
of demo-bank's own headless OAuth walk. The example is now
`pnpm --filter demo-bank mcp:agent`, and the OAuth dance exists once.

- `scripts/mcp-oauth.ts` — the walk, lifted verbatim out of `mcp-e2e.ts`
  (discovery, DCR, PKCE S256, Maple's login, the door's consent page, code
  exchange), with the client name parameterised.
- `scripts/mcp-stock-agent.ts` — what mcp-agent actually added: connect,
  take `tools/list` verbatim, hand it to `streamText`, and assert the
  receipt law. Still zero `@vendoai` imports.
- The consent page's theme assertion had drifted to `#0A7CFF`, a colour
  Maple has never used; it now reads the accent from `.vendo/theme.json`.
