+++ examples/demo-bank/package.json
@@ -10,6 +10,7 @@
     "build": "next build",
     "start": "next start",
     "mcp:e2e": "tsx scripts/mcp-e2e.ts",
+    "mcp:agent": "tsx scripts/mcp-stock-agent.ts",
     "seed:console": "tsx scripts/seed-console.ts",
     "lint": "eslint",
     "test": "vitest run"
