--- examples/mcp-agent/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@vendoai-examples/mcp-agent",
-  "version": "0.0.1",
-  "private": true,
-  "type": "module",
-  "description": "A stock AI SDK agent that reaches a Vendo-equipped product through the MCP door — zero Vendo imports. The doors proof: `vendo_make` called from outside, the receipt narrated by the calling agent, the screen landing on the product's own page.",
-  "engines": {
-    "node": ">=20"
-  },
-  "scripts": {
-    "agent": "node agent.mjs"
-  },
-  "dependencies": {
-    "@ai-sdk/anthropic": "3.0.12",
-    "@modelcontextprotocol/sdk": "^1.29.0",
-    "ai": "6.0.28"
-  }
-}
